refactor: optimize folders structure (created utils folder)
